+++ src/main/java/org/sopt/carena/member/exception/jwt/InvalidTokenException.java
@@ -1,7 +1,7 @@
-package org.sopt.carena.member.appliacation.exception.jwt;
+package org.sopt.carena.member.exception.jwt;
 
 import org.sopt.carena.global.exception.BaseException;
-import org.sopt.carena.member.appliacation.code.MemberErrorCode;
+import org.sopt.carena.member.exception.code.MemberErrorCode;
 
 public class InvalidTokenException extends BaseException {
